ai-systems/web/backend/requirements.txt
fastapi==0.100.0
uvicorn==0.24.0
python-dotenv==1.0.0
python-gitlab==3.15.0
pydantic==1.10.8
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
pika==1.3.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiohttp==3.9.1
aioredis==2.0.1
jinja2==3.1.2
requests==2.31.0
aiofiles==23.2.1
fastapi-cors==0.0.6
websockets==10.4
prompt-toolkit==3.0.38
prometheus-client==0.17.1
pytz==2023.3
markdown==3.4.4
loguru==0.7.0
multidict==6.0.4
yarl==1.9.2
fastapi-websocket-pubsub==0.3.1
fastapi-websocket-rpc==0.1.24
